fix: historical task filtering
